FREEMARKER-144 Proof Of Concept for providing DataFrames
